+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/page/BaseFireCloudPage.scala
@@ -1,5 +1,6 @@
 package org.broadinstitute.dsde.firecloud.page
 
+import org.broadinstitute.dsde.firecloud.component.{Link, TestId}
 import org.broadinstitute.dsde.firecloud.page.library.DataLibraryPage
 import org.broadinstitute.dsde.firecloud.page.methodrepo.MethodRepoPage
 import org.broadinstitute.dsde.firecloud.page.workspaces.WorkspaceListPage
@@ -8,17 +9,17 @@ import org.openqa.selenium.WebDriver
 abstract class BaseFireCloudPage(implicit webDriver: WebDriver) extends AuthenticatedPage {
 
   def goToWorkspaces(): WorkspaceListPage = {
-    click on testId("workspace-nav-link")
+    Link(TestId("workspace-nav-link")).doClick()
     await ready new WorkspaceListPage
   }
 
   def goToDataLibrary(): DataLibraryPage = {
-    click on testId("library-nav-link")
+    Link(TestId("library-nav-link")).doClick()
     await ready new DataLibraryPage
   }
 
   def goToMethodRepository(): MethodRepoPage = {
-    click on testId("method-repo-nav-link")
+    Link(TestId("method-repo-nav-link")).doClick()
     await ready new MethodRepoPage()
   }
 }
