+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/page/workspaces/methodconfigs/WorkspaceMethodConfigDetailsPage.scala
@@ -104,6 +104,7 @@ class WorkspaceMethodConfigDetailsPage(namespace: String, name: String, methodCo
 
   // TODO This is a very weak check to deterimine if page is ready
   def isLoaded: Boolean = {
+    await condition (webDriver.getCurrentUrl == url)
     openLaunchAnalysisModalButton.isVisible
   }
